+++ nodejs/src/clientchannel/service.ts
@@ -16,8 +16,10 @@ import {
   UntypedServiceImplementation,
 } from "@grpc/grpc-js";
 import { PublishEventEnvelope } from "./event";
+import { GetEventRequest } from "./get_event";
 import { GetEventsFromStartRequest } from "./get_events_from_start";
 import { GetStreamRequest, GetStreamResponse } from "./get_stream";
+import { IntroduceGdprOnEventFieldRequest, IntroduceGdprOnEventFieldResponse } from "./introduce_gdpr_on_event_field";
 import { IntroduceGdprOnFieldRequest, IntroduceGdprOnFieldResponse } from "./introduce_gdpr_on_field";
 import { InvalidateGdprRequest, InvalidateGdprResponse } from "./invalidate_gdpr";
 import { PublishRequest, PublishResponse } from "./publish";
@@ -55,6 +57,15 @@ export const ServiceService = {
     responseSerialize: (value: GetStreamResponse) => Buffer.from(GetStreamResponse.encode(value).finish()),
     responseDeserialize: (value: Buffer) => GetStreamResponse.decode(value),
   },
+  getEvent: {
+    path: "/clientchannel.Service/GetEvent",
+    requestStream: false,
+    responseStream: false,
+    requestSerialize: (value: GetEventRequest) => Buffer.from(GetEventRequest.encode(value).finish()),
+    requestDeserialize: (value: Buffer) => GetEventRequest.decode(value),
+    responseSerialize: (value: PublishEventEnvelope) => Buffer.from(PublishEventEnvelope.encode(value).finish()),
+    responseDeserialize: (value: Buffer) => PublishEventEnvelope.decode(value),
+  },
   getEventsFromStart: {
     path: "/clientchannel.Service/GetEventsFromStart",
     requestStream: false,
@@ -75,6 +86,17 @@ export const ServiceService = {
     responseSerialize: (value: IntroduceGdprOnFieldResponse) =>
       Buffer.from(IntroduceGdprOnFieldResponse.encode(value).finish()),
     responseDeserialize: (value: Buffer) => IntroduceGdprOnFieldResponse.decode(value),
+  },
+  introduceGdprOnEventField: {
+    path: "/clientchannel.Service/IntroduceGdprOnEventField",
+    requestStream: false,
+    responseStream: false,
+    requestSerialize: (value: IntroduceGdprOnEventFieldRequest) =>
+      Buffer.from(IntroduceGdprOnEventFieldRequest.encode(value).finish()),
+    requestDeserialize: (value: Buffer) => IntroduceGdprOnEventFieldRequest.decode(value),
+    responseSerialize: (value: IntroduceGdprOnEventFieldResponse) =>
+      Buffer.from(IntroduceGdprOnEventFieldResponse.encode(value).finish()),
+    responseDeserialize: (value: Buffer) => IntroduceGdprOnEventFieldResponse.decode(value),
   },
   invalidateGdpr: {
     path: "/clientchannel.Service/InvalidateGdpr",
@@ -100,8 +122,10 @@ export interface ServiceServer extends UntypedServiceImplementation {
   subscribe: handleBidiStreamingCall<SubscribeRequest, SubscribeResponse>;
   publish: handleUnaryCall<PublishRequest, PublishResponse>;
   getStream: handleUnaryCall<GetStreamRequest, GetStreamResponse>;
+  getEvent: handleUnaryCall<GetEventRequest, PublishEventEnvelope>;
   getEventsFromStart: handleServerStreamingCall<GetEventsFromStartRequest, PublishEventEnvelope>;
   introduceGdprOnField: handleUnaryCall<IntroduceGdprOnFieldRequest, IntroduceGdprOnFieldResponse>;
+  introduceGdprOnEventField: handleUnaryCall<IntroduceGdprOnEventFieldRequest, IntroduceGdprOnEventFieldResponse>;
   invalidateGdpr: handleUnaryCall<InvalidateGdprRequest, InvalidateGdprResponse>;
   snapshot: handleUnaryCall<SnapshotRequest, SnapshotResponse>;
 }
@@ -143,6 +167,21 @@ export interface ServiceClient extends Client {
     options: Partial<CallOptions>,
     callback: (error: ServiceError | null, response: GetStreamResponse) => void,
   ): ClientUnaryCall;
+  getEvent(
+    request: GetEventRequest,
+    callback: (error: ServiceError | null, response: PublishEventEnvelope) => void,
+  ): ClientUnaryCall;
+  getEvent(
+    request: GetEventRequest,
+    metadata: Metadata,
+    callback: (error: ServiceError | null, response: PublishEventEnvelope) => void,
+  ): ClientUnaryCall;
+  getEvent(
+    request: GetEventRequest,
+    metadata: Metadata,
+    options: Partial<CallOptions>,
+    callback: (error: ServiceError | null, response: PublishEventEnvelope) => void,
+  ): ClientUnaryCall;
   getEventsFromStart(
     request: GetEventsFromStartRequest,
     options?: Partial<CallOptions>,
@@ -166,6 +205,21 @@ export interface ServiceClient extends Client {
     metadata: Metadata,
     options: Partial<CallOptions>,
     callback: (error: ServiceError | null, response: IntroduceGdprOnFieldResponse) => void,
+  ): ClientUnaryCall;
+  introduceGdprOnEventField(
+    request: IntroduceGdprOnEventFieldRequest,
+    callback: (error: ServiceError | null, response: IntroduceGdprOnEventFieldResponse) => void,
+  ): ClientUnaryCall;
+  introduceGdprOnEventField(
+    request: IntroduceGdprOnEventFieldRequest,
+    metadata: Metadata,
+    callback: (error: ServiceError | null, response: IntroduceGdprOnEventFieldResponse) => void,
+  ): ClientUnaryCall;
+  introduceGdprOnEventField(
+    request: IntroduceGdprOnEventFieldRequest,
+    metadata: Metadata,
+    options: Partial<CallOptions>,
+    callback: (error: ServiceError | null, response: IntroduceGdprOnEventFieldResponse) => void,
   ): ClientUnaryCall;
   invalidateGdpr(
     request: InvalidateGdprRequest,